Relocate useAuthState to hooks
